feature/CMS-154 Unify validation logic in shared package, fix PFD defaults & clean architecture violations (Product Field Module)
