+++ backend/src/main/java/com/backend/application/dto/request/CreatePageSlotRequest.java
@@ -1,12 +1,13 @@
 package com.backend.application.dto.request;
 
-import com.backend.presentation.validation.SlotName;
-import com.backend.presentation.validation.Uid;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-
 import static com.backend.shared.constants.ValidationConstants.SLOT_NAME_MAX_LENGTH;
 import static com.backend.shared.constants.ValidationConstants.UID_MAX_LENGTH;
+
+import com.backend.shared.validation.SlotName;
+import com.backend.shared.validation.Uid;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
 import lombok.Data;
 
 @Data
