+++ backend/src/main/java/com/backend/application/dto/request/CreateProductFieldRequest.java
@@ -3,9 +3,9 @@ package com.backend.application.dto.request;
 import java.util.Map;
 
 import com.backend.domain.enums.ProductFieldType;
-import com.backend.presentation.validation.Code;
-import com.backend.presentation.validation.Uid;
 import com.backend.shared.constants.ValidationConstants;
+import com.backend.shared.validation.Code;
+import com.backend.shared.validation.Uid;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -16,24 +16,19 @@ import jakarta.validation.constraints.Size;
  * Uses project validation annotations for consistent pattern enforcement.
  */
 public record CreateProductFieldRequest(
-    @Uid(maxLength = ValidationConstants.CODE_MAX_LENGTH)
-    String uid,
+        @Uid(maxLength = ValidationConstants.UID_MAX_LENGTH) String uid,
 
-    @Code(maxLength = ValidationConstants.CODE_MAX_LENGTH)
-    String code,
+        @Code(maxLength = ValidationConstants.CODE_MAX_LENGTH) String code,
 
-    @NotBlank(message = "{validation.field.name.required}")
-    @Size(max = ValidationConstants.ATTRIBUTE_NAME_MAX_LENGTH, message = "{validation.field.name.maxLength}")
-    String name,
+        @NotBlank(message = "{validation.field.name.required}") @Size(max = ValidationConstants.ATTRIBUTE_NAME_MAX_LENGTH, message = "{validation.field.name.maxLength}") String name,
 
-    @NotNull(message = "{validation.field.fieldType.required}")
-    ProductFieldType fieldType,
+        @NotNull(message = "{validation.field.fieldType.required}") ProductFieldType fieldType,
 
-    Boolean isRequired,
-    Boolean isVisibleInList,
-    Integer sortOrder,
-    String defaultValue,
-    Map<String, Object> validationConfig) {
+        Boolean isRequired,
+        Boolean isVisibleInList,
+        Integer sortOrder,
+        String defaultValue,
+        Map<String, Object> validationConfig) {
 
     /**
      * Compact constructor for input sanitization.
