+++ backend/src/main/java/com/backend/domain/entity/ProductFieldDefinition.java
@@ -1,27 +1,18 @@
 package com.backend.domain.entity;
 
-import java.time.LocalDateTime;
-import java.util.UUID;
-
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
 import com.backend.domain.enums.ProductFieldType;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.PrePersist;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import lombok.experimental.SuperBuilder;
 
 /**
  * Entity representing a global product field definition.
@@ -33,18 +24,8 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
-public class ProductFieldDefinition {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @Column(nullable = false, unique = true, length = 36)
-  private String uuid;
-
-  @Column(nullable = false, unique = true, length = 50)
-  private String uid;
+@SuperBuilder
+public class ProductFieldDefinition extends BaseEntity {
 
   @Column(nullable = false, unique = true, length = 50)
   private String code;
@@ -73,19 +54,4 @@ public class ProductFieldDefinition {
 
   @Column(name = "validation_config", columnDefinition = "JSON")
   private String validationConfig;
-
-  @CreationTimestamp
-  @Column(name = "created_at", nullable = false, updatable = false)
-  private LocalDateTime createdAt;
-
-  @UpdateTimestamp
-  @Column(name = "updated_at", nullable = false)
-  private LocalDateTime updatedAt;
-
-  @PrePersist
-  public void prePersist() {
-    if (this.uuid == null) {
-      this.uuid = UUID.randomUUID().toString();
-    }
-  }
 }
