+++ backend/src/main/java/com/backend/presentation/dto/request/AttributeDefinitionCreateRequest.java
@@ -1,24 +1,21 @@
 package com.backend.presentation.dto.request;
 
+import static com.backend.shared.constants.ValidationConstants.ATTRIBUTE_CODE_MAX_LENGTH;
+import static com.backend.shared.constants.ValidationConstants.ATTRIBUTE_NAME_MAX_LENGTH;
+
 import com.backend.domain.enums.ProductFieldType;
-import com.backend.presentation.validation.Code;
+import com.backend.shared.validation.Code;
+
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
-import static com.backend.shared.constants.ValidationConstants.ATTRIBUTE_CODE_MAX_LENGTH;
-import static com.backend.shared.constants.ValidationConstants.ATTRIBUTE_NAME_MAX_LENGTH;
-
 public record AttributeDefinitionCreateRequest(
-        @Code(maxLength = ATTRIBUTE_CODE_MAX_LENGTH)
-        String code,
+        @Code(maxLength = ATTRIBUTE_CODE_MAX_LENGTH) String code,
 
-        @NotBlank(message = "validation.attribute.name.required")
-        @Size(max = ATTRIBUTE_NAME_MAX_LENGTH, message = "validation.attribute.name.size")
-        String name,
+        @NotBlank(message = "validation.attribute.name.required") @Size(max = ATTRIBUTE_NAME_MAX_LENGTH, message = "validation.attribute.name.size") String name,
 
-        @NotNull(message = "validation.attribute.fieldType.required")
-        ProductFieldType fieldType,
+        @NotNull(message = "validation.attribute.fieldType.required") ProductFieldType fieldType,
 
         Boolean isRequired,
 
@@ -26,13 +23,15 @@ public record AttributeDefinitionCreateRequest(
 
         Integer sortOrder,
 
-        String validationConfig
-) {
+        String validationConfig) {
     public AttributeDefinitionCreateRequest {
         code = code != null ? code.trim() : null;
         name = name != null ? name.trim() : null;
-        if (isRequired == null) isRequired = false;
-        if (isSearchable == null) isSearchable = false;
-        if (sortOrder == null) sortOrder = 0;
+        if (isRequired == null)
+            isRequired = false;
+        if (isSearchable == null)
+            isSearchable = false;
+        if (sortOrder == null)
+            sortOrder = 0;
     }
 }
