+++ backend/src/main/java/com/backend/presentation/dto/request/CategoryCompositeRequest.java
@@ -1,17 +1,17 @@
 package com.backend.presentation.dto.request;
 
-import com.backend.domain.enums.Language;
-import com.backend.presentation.validation.CategoryCode;
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotEmpty;
+import static com.backend.shared.constants.ValidationConstants.CATEGORY_CODE_MAX_LENGTH;
 
 import java.util.Map;
 
-import static com.backend.shared.constants.ValidationConstants.CATEGORY_CODE_MAX_LENGTH;
+import com.backend.domain.enums.Language;
+import com.backend.shared.validation.CategoryCode;
+
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotEmpty;
 
 public record CategoryCompositeRequest(
-        @CategoryCode(maxLength = CATEGORY_CODE_MAX_LENGTH)
-        String code,
+        @CategoryCode(maxLength = CATEGORY_CODE_MAX_LENGTH) String code,
 
         Long parentId,
 
@@ -19,13 +19,12 @@ public record CategoryCompositeRequest(
 
         Boolean isVisible,
 
-        @NotEmpty(message = "validation.category.translations.required")
-        @Valid
-        Map<Language, CategoryI18nRequest> translations
-) {
+        @NotEmpty(message = "validation.category.translations.required") @Valid Map<Language, CategoryI18nRequest> translations) {
     public CategoryCompositeRequest {
         code = code != null ? code.trim() : null;
-        if (sortOrder == null) sortOrder = 0;
-        if (isVisible == null) isVisible = true;
+        if (sortOrder == null)
+            sortOrder = 0;
+        if (isVisible == null)
+            isVisible = true;
     }
 }
