+++ backend/src/main/java/com/backend/shared/validation/Uid.java
@@ -1,0 +1,29 @@
+package com.backend.shared.validation;
+
+import static com.backend.shared.constants.ValidationConstants.MSG_UID_PATTERN;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+
+@Documented
+@Constraint(validatedBy = UidValidator.class)
+@Target({ ElementType.FIELD, ElementType.PARAMETER, ElementType.METHOD })
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Uid {
+
+  String message() default MSG_UID_PATTERN;
+
+  Class<?>[] groups() default {};
+
+  Class<? extends Payload>[] payload() default {};
+
+  boolean required() default true;
+
+  int maxLength() default 100;
+}
