+++ backend/src/main/resources/i18n/messages_en.properties
@@ -397,7 +397,7 @@ product.field.delete.success=Product field deleted successfully
 product.field.not.found=Product field not found: {0}
 product.field.get.error=Error retrieving product field: {0}
 product.field.list.error=Error listing product fields: {0}
-error.max.field.limit=Maximum field limit exceeded
+product.field.max.limit=Maximum field limit exceeded
 
 # Category Messages
 category.create.success=Category created successfully
