+++ backend/src/main/resources/i18n/messages_tr.properties
@@ -62,7 +62,7 @@ error.data.duplicate=Bu veri zaten mevcut
 error.data.integrity=Veri doğrulama hatası
 error.validation.constraint=Veri gerekli kısıtlamaları karşılamıyor
 error.runtime=Beklenmeyen bir hata oluştu
-error.max.field.limit=Maksimum alan limiti aşıldı
+product.field.max.limit=Maksimum alan limiti aşıldı
 
 # Kimlik Doğrulama Mesajları
 auth.platform.login.success=Platform yönetici girişi başarılı
